Decouple player from camera
Make it so the camera and player is not a single scene. We can now dynamically swap between freeform camera movement and attached to the player where camera follows the player.

This is done so we can switch UI edit style in controller scene not the whole playerscene where it would become issue if player is dead.

Player can respawn in this game during mission even though camera can be moved during death that is done a lot later when we start development on gameplay side.